chore: move support bundle generation code to machinery
This should allow using it in 3rd party projects without the need to
pull the main Talos repo.
